backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.1
pypdf==4.2.0
